+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MV-187 - Upload vector image
 - VIMEET-1935 - Generate promo codes
 - VIMEET-2018 - Export promo codes group
+- VIMEET-2017 - Update promo codes group
 
 ## [2.10.0] - 2019-07-05
 ### Added